requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
cachetools==5.5.2
    # via google-auth
certifi==2025.7.9
    # via requests
charset-normalizer==3.4.2
    # via requests
coloredlogs==15.0.1
    # via onnxruntime
datasketch==1.6.5
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flatbuffers==25.2.10
    # via onnxruntime
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.176.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.1
    # via
    #   google-api-core
    #   grpcio-status
    #   grpcio-tools
grpcio-status==1.71.2
    # via google-api-core
grpcio-tools==1.71.2
    # via -r requirements.in
hf-xet==1.1.5
    # via huggingface-hub
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
huggingface-hub==0.33.4
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via requests
jieba==0.42.1
    # via -r requirements.in
jinja2==3.1.6
    # via torch
joblib==1.5.1
    # via scikit-learn
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   -r requirements.in
    #   datasketch
    #   onnx
    #   onnxruntime
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
onnx==1.18.0
    # via skl2onnx
onnxruntime==1.22.1
    # via -r requirements.in
packaging==25.0
    # via
    #   huggingface-hub
    #   onnxruntime
    #   transformers
pandas==2.3.1
    # via -r requirements.in
pillow==11.3.0
    # via sentence-transformers
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   grpcio-tools
    #   onnx
    #   onnxruntime
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via google-generativeai
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   google-api-core
    #   huggingface-hub
    #   transformers
rsa==4.9.1
    # via google-auth
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via
    #   -r requirements.in
    #   sentence-transformers
    #   skl2onnx
scipy==1.16.0
    # via
    #   datasketch
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.0.0
    # via -r requirements.in
setuptools==80.9.0
    # via
    #   grpcio-tools
    #   torch
    #   triton
six==1.17.0
    # via python-dateutil
skl2onnx==1.19.1
    # via -r requirements.in
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
tabulate==0.9.0
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   google-generativeai
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.53.2
    # via sentence-transformers
triton==3.3.1
    # via torch
typing-extensions==4.14.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   onnx
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
